+++ programs/dcarbon-contract/src/state/config.rs
@@ -1,10 +1,8 @@
 use anchor_lang::prelude::*;
-use anchor_lang::system_program::{create_account, CreateAccount};
 
 use crate::error::DCarbonError;
-use crate::ID;
 use crate::instructions::ConfigArgs;
-use crate::state::{BpfWriter, DeviceLimit};
+use crate::state::DeviceLimit;
 
 #[account]
 #[derive(Debug, InitSpace)]
@@ -15,14 +13,14 @@ pub struct ContractConfig {
     pub mint: Pubkey,
     #[max_len(100)]
     pub minting_limits: Vec<DeviceLimit>,
-    pub governance_amount: u64,
+    pub governance_amount: f64,
 }
 
 impl ContractConfig {
     pub const PREFIX_SEED: &'static [u8] = b"contract_config";
 
     pub fn validate(&self, config_args: ConfigArgs) -> Result<()> {
-        if config_args.rate <= 0.0 || config_args.minting_fee <= 0.0 || config_args.governance_amount <= 0 {
+        if config_args.rate <= 0.0 || config_args.minting_fee <= 0.0 || config_args.governance_amount <= 0.0 {
             return Err(DCarbonError::InitArgsInvalid.into());
         }
         Ok(())
